requirements.txt
flask
streamlit
plotly
groq
python-dotenv
twilio
pandas
numpy
